src/requirements.txt
fastapi
uvicorn
requests
pytest
openai
python-dotenv
azure-identity
azure-keyvault-secrets
starlette